fix(e2e): unskip and modernize BYOK/wallet hosted tests, sessions result count
